Updated Sign Up page to add lat & long to database
